rename chain configs files
