--- web/src/config/chainConfigs/dusk.ts
@@ -1,115 +1,0 @@
-import type { EvmChainInfo, EvmChains, IbcChainInfo, IbcChains } from ".";
-
-const CelestiaChainInfo: IbcChainInfo = {
-  // Chain-id of the celestia chain.
-  chainId: "mocha-4",
-  // The name of the chain to be displayed to the user.
-  chainName: "Celestia Mocha-4",
-  // RPC endpoint of the chain
-  rpc: "wss://rpc-mocha.pops.one",
-  // REST endpoint of the chain.
-  rest: "https://api-mocha.pops.one",
-  // Staking coin information
-  stakeCurrency: {
-    // Coin denomination to be displayed to the user.
-    coinDenom: "TIA",
-    // Actual denom (i.e. uatom, uscrt) used by the blockchain.
-    coinMinimalDenom: "utia",
-    // # of decimal points to convert minimal denomination to user-facing denomination.
-    coinDecimals: 6,
-    // (Optional) Keplr can show the fiat value of the coin if a coingecko id is provided.
-    // You can get id from https://api.coingecko.com/api/v3/coins/list if it is listed.
-    // coinGeckoId: ""
-  },
-  // (Optional) If you have a wallet webpage used to stake the coin then provide the url to the website in `walletUrlForStaking`.
-  // The 'stake' button in Keplr extension will link to the webpage.
-  // walletUrlForStaking: "",
-  // The BIP44 path.
-  bip44: {
-    // You can only set the coin type of BIP44.
-    // 'Purpose' is fixed to 44.
-    coinType: 118,
-  },
-  // The address prefix of the chain.
-  bech32Config: {
-    bech32PrefixAccAddr: "celestia",
-    bech32PrefixAccPub: "celestiapub",
-    bech32PrefixConsAddr: "celestiavalcons",
-    bech32PrefixConsPub: "celestiavalconspub",
-    bech32PrefixValAddr: "celestiavaloper",
-    bech32PrefixValPub: "celestiavaloperpub",
-  },
-  // List of all coin/tokens used in this chain.
-  currencies: [
-    {
-      // Coin denomination to be displayed to the user.
-      coinDenom: "TIA",
-      // Actual denom (i.e. uatom, uscrt) used by the blockchain.
-      coinMinimalDenom: "utia",
-      // # of decimal points to convert minimal denomination to user-facing denomination.
-      coinDecimals: 6,
-      // (Optional) Keplr can show the fiat value of the coin if a coingecko id is provided.
-      // You can get id from https://api.coingecko.com/api/v3/coins/list if it is listed.
-      // coinGeckoId: ""
-      ibcChannel: "channel-128",
-      sequencerBridgeAccount: "astria17hvahh8lcas4fhl5urqjnhfqwhffkddaw034lh",
-      iconClass: "i-celestia",
-    },
-  ],
-  // List of coin/tokens used as a fee token in this chain.
-  feeCurrencies: [
-    {
-      // Coin denomination to be displayed to the user.
-      coinDenom: "TIA",
-      // Actual denom (i.e. nria, uscrt) used by the blockchain.
-      coinMinimalDenom: "utia",
-      // # of decimal points to convert minimal denomination to user-facing denomination.
-      coinDecimals: 6,
-      // (Optional) Keplr can show the fiat value of the coin if a coingecko id is provided.
-      // You can get id from https://api.coingecko.com/api/v3/coins/list if it is listed.
-      // coinGeckoId: ""
-      // (Optional) This is used to set the fee of the transaction.
-      // If this field is not provided and suggesting chain is not natively integrated, Keplr extension will set the Keplr default gas price (low: 0.01, average: 0.025, high: 0.04).
-      // Currently, Keplr doesn't support dynamic calculation of the gas prices based on on-chain data.
-      // Make sure that the gas prices are higher than the minimum gas prices accepted by chain validators and RPC/REST endpoint.
-      gasPriceStep: {
-        low: 0.01,
-        average: 0.02,
-        high: 0.1,
-      },
-    },
-  ],
-  iconClass: "i-celestia",
-};
-
-export const ibcChains: IbcChains = {
-  "Celestia Mocha-4": CelestiaChainInfo,
-};
-
-const FlameChainInfo: EvmChainInfo = {
-  chainId: 912559,
-  chainName: "Flame Dusk-10",
-  rpcUrls: ["https://rpc.evm.dusk-10.devnet.astria.org"],
-  currencies: [
-    {
-      coinDenom: "RIA",
-      coinMinimalDenom: "uria",
-      coinDecimals: 18,
-      iconClass: "i-celestia",
-    },
-    {
-      coinDenom: "TIA",
-      coinMinimalDenom: "utia",
-      coinDecimals: 6,
-      evmWithdrawerContractAddress:
-        "0xFc83F6A786728F448481B7D7d5C0659A92a62C4d",
-      iconClass: "i-celestia",
-    },
-  ],
-  // TODO - import icon
-  iconClass: "i-flame",
-};
-
-export const evmChains: EvmChains = {
-  "Flame Dusk-10": FlameChainInfo,
-};
